requirements.txt
asgiref==3.3.1
backcall==0.2.0
bleach==3.2.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.2
Django==3.1.5
django-configurations==2.2
django-cors-headers==3.6.0
django-cuser==2017.3.16
django-modeltranslation==0.16.1
django-tinymce==3.2.0
django-user-agents==0.4.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
drf-api-logger==0.0.9
flake8==3.8.4
h11==0.12.0
httpcore==0.12.2
httpx==0.16.1
idna==2.10
ipython==7.19.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.8
parso==0.8.1
phonenumbers==8.12.16
pickleshare==0.7.5
Pillow==8.1.0
prompt-toolkit==3.0.10
psycopg2==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.7.4
PyJWT==1.7.1
pyparsing==2.4.7
python-dotenv==0.15.0
python-memcached==1.59
pytz==2020.5
requests==2.25.1
rfc3986==1.4.0
six==1.15.0
sniffio==1.2.0
sqlparse==0.4.1
traitlets==5.0.5
ua-parser==0.10.0
uritemplate==3.0.1
urllib3==1.26.2
user-agents==2.2.0
wcwidth==0.2.5
webencodings==0.5.1
